+++ packages/std/src/query/bank.rs
@@ -7,6 +7,11 @@ use crate::Coin;
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
 pub enum BankQuery {
+    /// This calls into the native bank module for querying the total supply of one denomination.
+    /// It does the same as the SupplyOf call in Cosmos SDK's RPC API.
+    /// Return value is of type SupplyResponse.
+    #[cfg(feature = "cosmwasm_1_1")]
+    Supply { denom: String },
     /// This calls into the native bank module for one denomination
     /// Return value is BalanceResponse
     Balance { address: String, denom: String },
@@ -14,6 +19,16 @@ pub enum BankQuery {
     /// Note that this may be much more expensive than Balance and should be avoided if possible.
     /// Return value is AllBalanceResponse.
     AllBalances { address: String },
+}
+
+#[cfg(feature = "cosmwasm_1_1")]
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+#[non_exhaustive]
+pub struct SupplyResponse {
+    /// Always returns a Coin with the requested denom.
+    /// This will be of zero amount if the denom does not exist.
+    pub amount: Coin,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]