+++ packages/std/src/query/mod.rs
@@ -10,6 +10,8 @@ mod ibc;
 mod staking;
 mod wasm;
 
+#[cfg(feature = "cosmwasm_1_1")]
+pub use bank::SupplyResponse;
 pub use bank::{AllBalanceResponse, BalanceResponse, BankQuery};
 #[cfg(feature = "stargate")]
 pub use ibc::{ChannelResponse, IbcQuery, ListChannelsResponse, PortIdResponse};